Problem 3 refactored
