+++ src/test/java/Tests.java
@@ -1,6 +1,7 @@
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
+import po.StringExercises;
 
 public class Tests {
     private StringExercises stringExercises;
@@ -91,31 +92,30 @@ public class Tests {
 
     // Problem 3 Tests
     @Test
-    public void reverseWordTest(){
-        String actual=stringExercises.reverseWord("home");
-        String expected="emoh";
+    public void reverseWordTest() {
+        String actual = stringExercises.reverseWord("home");
+        String expected = "emoh";
         Assert.assertEquals(expected, actual);
     }
 
     @Test
-    public void addDashTest(){
-        String actual=stringExercises.addDash("water");
-        String expected="w-a-t-e-r";
+    public void addDashTest() {
+        String actual = stringExercises.addDash("water");
+        String expected = "w-a-t-e-r";
         Assert.assertEquals(expected, actual);
     }
 
     @Test
-    public void getVowelsTest(){
-        String actual=stringExercises.getVowels("window");
-        String expected="io";
+    public void getVowelsTest() {
+        String actual = stringExercises.getVowels("window");
+        String expected = "io";
         Assert.assertEquals(expected, actual);
     }
 
     @Test
     public void reverseVowelsTest() {
-        String actual = stringExercises.reverseVowels("MILk", "sUgar");
+        String actual = stringExercises.reverseVowels("MIlk", "sUgar");
         String expected = "a-u-i";
         Assert.assertEquals(expected, actual);
     }
-
 }
